pinctrl: Kconfig: sort pinctrl config options to prevent future clutter
This originally started out as
     "pinctrl: Kconfig: reorder to keep Rockchip options together"
and tried to keep the Rockchip-related config options together.

However, we now rewrite all chip-specific driver selections to start
with CONFIG_PINCTRL_ (with the inadvertent changes to related
Makefiles) and sort those alphabetically. And as this already means
touching most of the file, we also reformat the help text to not exceed
80 characters (but make full use of those 80 characters).

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>

Series-version: 3
Series-changes: 2
- sort options alphabetically and use a common naming (i.e. always
  start with CONFIG_PINCTRL_) for chip-specific drivers

Series-changes: 3
- sed lingering old options in defconfig to conform with the new
  naming scheme
- sed the ROCKCHIP_*_PINCTRL (which started this in the first place)
  into PINCTRL_ROCKCHIP_* for consistency with the new naming scheme