rockchip: mkimage: rksd: pad SD/MMC images to a full blocksize
While not required for correctness, we pad the SD/MMC images to a full
blocksize (i.e. 512 bytes). This is to keep things consistent with all
operations being performed on blocksize units.

Note that a case could be made to choose a 2K alignment (as init_size
needs to be a multiple of 4 blocks) and to keep even closer to rkspi,
but the 512 byte choice felt more natural as it's also the natural
blocksize of SD/MMC devices.

Series-changes: 2
- added missing commit message (with a bit of background to the
  in-line comment in the code).

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Acked-by: Simon Glass <sjg@chromium.org>