rockchip: defconfig: puma-rk3399: enable RK3399 efuse driver
With everything in place (i.e. the new efuse driver, the clk-support
for the non-secure efuse block, and the board-specific functions to
derive 'serial#' from the cpu-id within the efuses), enable this in
the RK3399-Q7 defconfig.

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>

Series-cc: Klaus Goger <klaus.goger@theobroma-systems.com>
Series-cc: Kever Yang <kever.yang@rock-chips.com>
Series-cc: Simon Glass <sjg@chromium.org>

Cover-letter:
rockchip: Add a read-only efuse driver for the RK3399 (and derive serial# from fuses on the RK3399-Q7)

As the RK3399-Q7 (Puma) differs from our previous modules in how we
can force an external boot (on the RK3399-Q7 this holds the eMMC and
SPI in reset, until an external signal is removed) through the 'BIOS
disable'-signal of the Qseven specification, we can't derive the
unique board id reliably from the eMMC's unique id.

Instead, for the RK3399-Q7, we rely on the efuse block of the RK3399
and use the (assumed to be unique) cpu_id region.

This patch-series adds the required infrastructure and switches the
RK3399-Q7 (puma_rk3399) over to this new infrastructure:
 - adds a rockchip-efuse driver as a misc-device (and permits requests
   for the clock gate listed in the DTS for the efuse node)
 - ports the CRC32-based 'serial#' derivation over from Linux
 - exposes the complete cpu_id field through 'cpuid#'
END

Series-version: 2