rockchip: board: puma_rk3399: derive ethaddr from cpuid
Generate a MAC address based on the cpuid available in the efuse
block: Use the first 6 byte of the cpuid's SHA256 hash and set the
locally administered bits. Also ensure that the multicast bit is
cleared.

The MAC address is only generated and set if there is no ethaddr
present in the saved environment.

Signed-off-by: Klaus Goger <klaus.goger@theobroma-systems.com>
Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>

Series-changes: 2
- added derivation of ethaddr from cpuid

Reviewed-by: Simon Glass <sjg@chromium.org>

Series-changes: 3
- uses uclass_get_device_by_driver() to ensure we don't pick up the wrong
  misc-device