rockchip: video: rk3399: add HDMI TX support on the RK3399
This commit enables the RK3399 HDMI TX, which is very similar to the
one found on the RK3288. The differences between the two SoCs (mainly
the input VOP selection) is abstracted away through the driverdata.

Note that the I2C communication for reading the EDID works well with
the default settings, but does not with the alternate settings used on
the RK3288... so this configuration aspect also is part of the
driverdata.

Having some sort of DTS-based configuration for the regulator
dependencies would be nice for the future, but for now we simply use
lists of regulator names (also via driverdata) that we probe.

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Series-cc: Klaus Goger <klaus.goger@theobroma-systems.com>

Cover-letter:
rockchip: video: rk3399: enable HDMI output

This series provides HDMI enablement for the RK3399 and will support
video console on the RK3399 either with VOP-lit or VOP-big:
 - pinctrl and clk support for the hdmi node
 - a refactoring of rk_vop.c and rk_hdmi.c to allow for the minor
   differences between the RK3288 and RK3399 VOP and HDMI blocks

To see the full series in the context of a working configuration (i.e.
including DTS, defconfig, etc. and on top of the MIPI enablement
change series from Rockchip), refer to
       https://git.theobroma-systems.com/puma-u-boot.git/log/?h=puma-hdmi-rfc

This has been tested (on top of Eric Gao's change series for MIPI-DSI)
using 'bmp display' and 'setenv stdout vidconsole' on the RK3399-Q7
with various HDMI monitors, both for VOP-lit and VOP-big (setting the
other one to 'disabled').
END