rockchip: dts: rk3399-puma: add /config/arm-trusted-firmware,reset-gpio property
With the ATF capable of accessing the FDT passed to the next stage,
we can specify configuration items for the ATF in the /config path.

This adds the arm-trusted-firmware,reset-gpio that conveys the number
of the GPIO used to reset the board (used, when a reboot is requested
from ATF via PSCI).

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>