rockchip: puma-rk3399: Enable vdd-log during bootup.
On the RK3399-Q7 "Puma" module VDD_LOG is generated by an external
regulator, which sets the voltage level to 900 mV, which is within
the allowed range and which works quite fine.

However, in specific use-cases we need to adjust VDD_LOG to
maintain stable operation or to reduce power consumption.
This adjustment can be done via pwm2.

This patch allows to tune the VDD_LOG voltage level
via DTS. To do so the following things are done:

 * Setup pin muxing for PWM2 (based on DTS settings)
 * Turn on the vdd_log regulator (based on DTS settings)

Reported-by: Assaf Agmon <assaf@r-go.io>
Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>