rockchip: rk3399-puma: Reset from SPL on error.
Every now an then we experience errors in SPL
(most likely "DRAM init failed").
In such cases, the system hangs and requires a power-cycle
from the user.

While a power-cycle might be a proper way to address
the symptom, we could also call do_reset() instead
of hanging. That would allow to "heal" spurious issues.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>